requirement.txt
autopep8==1.6.0
Pillow==9.0.0
pycodestyle==2.8.0
pyglet==1.5.21
toml==0.10.2
